Move read/write functions to their own files.
